Rename mock file, update related tests
