+++ source/frontend/src/features/mapSideBar/property/tabs/propertyDetailsManagement/activity/edit/InvoiceTotalsForm.test.tsx
@@ -1,11 +1,11 @@
 import { Formik } from 'formik';
 
 import { mockLookups } from '@/mocks/lookups.mock';
-import { getMockPropertyManagementActivity } from '@/mocks/PropertyManagementActivity.mock';
+import { getMockPropertyManagementActivity } from '@/mocks/managementActivity.mock';
+import { getMockManagementActivityInvoice } from '@/mocks/managementActivityInvoice.mock';
 import { lookupCodesSlice } from '@/store/slices/lookupCodes';
 import { act, render, RenderOptions, screen, userEvent } from '@/utils/test-utils';
 
-import { getMockManagementActivityInvoice } from '@/mocks/managementActivityInvoice.mock';
 import { InvoiceTotalsForm } from './InvoiceTotalsForm';
 import { PropertyActivityFormModel } from './models';
