+++ source/frontend/src/mocks/managementActivity.mock.ts
